% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_format.R
\name{easy_format}
\alias{easy_format}
\title{Fonction de formatage d'un nombre pour affichage}
\usage{
easy_format(variable, type_out, decimal = 0, suffix = NULL)
}
\arguments{
\item{variable}{Variable à formater}

\item{type_out}{Type de nombre en sortie : "pourcent", "milliers"}

\item{decimal}{Nombre de décimales attendues. 0 Par défaut}

\item{suffix}{Symbole éventiel à ajouter. AUcun par défaut}
}
\value{
Renvoit une variable formatée en pourcentage ou milliers, avec les bonnes règles typographiques
}
\description{
Fonction de formatage d'un nombre pour affichage
}
