% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jcan.R
\docType{data}
\name{jcan_colors}
\alias{jcan_colors}
\title{Palette de couleurs personnalisées (adaptées au thème personnalisé}
\format{
An object of class \code{character} of length 10.
}
\usage{
jcan_colors
}
\value{
La fonction permet d'utiliser la palette de couleurs personnalisées
}
\description{
Palette de couleurs personnalisées (adaptées au thème personnalisé
}
\keyword{datasets}
