% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jcan.R
\name{jcan_cols}
\alias{jcan_cols}
\title{Fonction pour extraire mes couleurs comme des hex codes}
\usage{
jcan_cols(...)
}
\arguments{
\item{...}{Character names of jcan_colors}
}
\description{
Fonction pour extraire mes couleurs comme des hex codes
}
