% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jcan.R
\name{jcan_pal}
\alias{jcan_pal}
\title{Fonction pour interpoler une palette custom}
\usage{
jcan_pal(palette = "principale", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Nom de la palette dans jcan_palettes}

\item{reverse}{Boolean pour inverser ou non la palette sélectionnée}

\item{...}{Arguements additionnels à passer dans colorRampPalette()}
}
\value{
Renvoi la palette sélectionné, et l'interpole pour avoir plus de couleurs si nécessaire
}
\description{
Fonction pour interpoler une palette custom
}
