% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marimekko.R
\name{marimekko_plot}
\alias{marimekko_plot}
\title{Marimekko graphique avec ggplot2 ersion fonctionnelle avec transformation en facteur si nécessaire ====}
\usage{
marimekko_plot(dataset, group, fill, switch = TRUE, nb_car = 15, ...)
}
\arguments{
\item{dataset}{Source de données}

\item{group}{Variable de l'axe X}

\item{fill}{Découpage sur l'axe Y}

\item{switch}{Pour inverser l'axe d'affichage des étiquettes}

\item{nb_car}{Nombre de caractères par ligne dans les strip}

\item{...}{Arguments complémentaires}
}
\value{
Génère un graphique type Marimekko, avec une variable de type facteur sur l'axe X, une varaible de type facteur pour décomposer chaque groupe. La largeur de la colonne est définie par les effectifs du groupe
}
\description{
Marimekko graphique avec ggplot2 ersion fonctionnelle avec transformation en facteur si nécessaire ====
}
