% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moy_mob.R
\name{moy_mob}
\alias{moy_mob}
\title{Moyenne mobile d'ordre N}
\usage{
moy_mob(var, n = 5)
}
\arguments{
\item{var}{Nom du vecteur sur lequel calculé la moyenne mobile}

\item{n}{Ordre de la moyenne mobile}
}
\value{
Permet de lisser une série de données par la méthode des moyennes mobiles
}
\description{
Moyenne mobile d'ordre N
}
