% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jcan.R
\name{scale_color_jcan}
\alias{scale_color_jcan}
\title{Constructeur de scale_color pour ggplot avec les palettes custom du packages}
\usage{
scale_color_jcan(palette = "principale", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Nom de la palette dans jcan_palettes}

\item{discrete}{Boolean pour indiquer si la palette sélectionnée s'applique à des données discrètes ou non}

\item{reverse}{Boolean pour inverser ou non la palette sélectionnée}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
Applique la palette custom sur un graphique ggplot
}
\description{
Constructeur de scale_color pour ggplot avec les palettes custom du packages
}
