% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jcan_marimekko.R
\name{theme_jcan_marimekko}
\alias{theme_jcan_marimekko}
\title{Thème personnalisé}
\usage{
theme_jcan_marimekko()
}
\value{
La fonction applique un thème à un graphique ggplot2
}
\description{
Thème personnalisé
}
