% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jcan.R
\name{jcan_cols}
\alias{jcan_cols}
\title{Fonction pour extraire mes couleurs comme des hex codes}
\usage{
jcan_cols(...)
}
\arguments{
\item{...}{Noms en toutes lettes pour les couleurs de ma palette custom jcan_colors}
}
\value{
La fonctionne transforme les noms de couleurs en HEX code
}
\description{
Fonction pour extraire mes couleurs comme des hex codes
}
