% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jcan.R
\name{jcan_pal}
\alias{jcan_pal}
\title{Fonction pour interpoler une palette custom}
\usage{
jcan_pal(palette = "principale", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in jcan_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette() comme l'inversion de la palette}
}
\description{
Fonction pour interpoler une palette custom
}
